Module Federation added
