+++ package.json
@@ -9,6 +9,8 @@
     "lint": "next lint"
   },
   "dependencies": {
+    "@module-federation/nextjs-mf": "^8.6.0",
+    "@module-federation/utilities": "1.7.3",
     "@project-sunbird/sunbird-questionset-editor-web-component": "^4.0.2",
     "async": "^3.2.5",
     "body-parser": "^1.20.2",
@@ -29,6 +31,7 @@
   },
   "devDependencies": {
     "eslint": "^8",
-    "eslint-config-next": "14.2.5"
+    "eslint-config-next": "14.2.5",
+    "webpack": "^5.76.2"
   }
 }
